remove my projects , remove types file ,  fixed link path in change password
