--- src/Pages/Employee/MyProjects/MyProjects.tsx
@@ -1,13 +1,0 @@
-import React from 'react'
-import Header from '../../../components/Header/Header';
-
-export default function MyProjects() {
-
-  
-  return (
-    <>
-      <Header />
-      <h1>my projects </h1>
-    </>
-  );
-}
